+++ src/IrreversibleUnimolecularReaction.h
@@ -66,7 +66,7 @@ namespace mesmer
 //X    void getPdtsCellDensityOfStates(std::vector<double> &cellDOS) ;
 
     // Return products
-    virtual int get_products(std::vector<Molecule *> &product) const
+    virtual size_t get_products(std::vector<Molecule *> &product) const
     {
       product.push_back(m_pdt1) ;
       if(m_pdt2){
@@ -80,7 +80,7 @@ namespace mesmer
       return 1;
     } ;
 
-    virtual int get_reactants(std::vector<Molecule *> &reactants) const
+    virtual size_t get_reactants(std::vector<Molecule *> &reactants) const
     {
       reactants.push_back(m_rct1);
       return 1;
